+++ app/Http/Controllers/CityController.php
@@ -30,12 +30,7 @@ class CityController extends Controller
         return view('cities.create', compact('governorates'));
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+
     public function store(Request $request)
     {
         $rules = [
@@ -47,7 +42,7 @@ class CityController extends Controller
         ];
         $this->validate($request, $rules, $message);
         $record = City::create($request->all());
-        flash()->success('<p style="text-align: center;font-weight: bolder">تــم اضــافة المدينة بنجــاح</p>');
+        flash()->success('تمت إضافة المدينة بنجاح');
         return redirect(route('cities.index'));
     }
 
@@ -97,9 +92,20 @@ class CityController extends Controller
      */
     public function destroy($id)
     {
-        $record = City::findOrFail($id);
+        $record = City::find($id);
+        if (!$record) {
+            return response()->json([
+                'status'  => 0,
+                'message' => 'تعذر الحصول على البيانات'
+            ]);
+        }
+
         $record->delete();
-        flash()->error('<p class="text-center" style="font-size:20px; font-weight:900;font-family:Arial" >تـــم الحــذف </p>');
-        return back();
+        return response()->json([
+                'status'  => 1,
+                'message' => 'تم الحذف بنجاح',
+                'id'      => $id
+            ]);
     }
+
 }
